version 0.24 - toggle for forms, popup with confirm, backend checks

+++ client/src/components/DeleteCategoryButton.js
@@ -1,12 +1,15 @@
 import { useMutation } from '@apollo/client'
+import { useConfirmDialog } from '../hooks/useConfirmDialog'
 import { DELETE_CATEGORY } from '../graphql/mutations'
 import { ALL_CATEGORIES } from '../graphql/queries'
 import useNotification from '../context/NotificationContext'
 import { FaTrashAlt } from 'react-icons/fa'
+import ConfirmDialog from './ConfirmDialog'
 import '../styles/components/CategoryForm.css'
 
 const DeleteCategoryButton = ({ id }) => {
   const { setNotification } = useNotification()
+  const { dialog, handleInputMessage, handleActionDialog } = useConfirmDialog()
 
   const [deleteItem] = useMutation(DELETE_CATEGORY, {
     variables: id,
@@ -23,15 +26,23 @@ const DeleteCategoryButton = ({ id }) => {
     },
   })
 
-  const triggerDeletion = id => {
-    deleteItem({ variables: { categoryId: id } })
+  const dialogConfirmation = confirm => {
+    if (confirm) {
+      handleActionDialog('', false)
+      deleteItem({ variables: { categoryId: id } })
+    } else {
+      handleActionDialog('', false)
+    }
   }
 
   return (
     <div>
       <FaTrashAlt
         className="deleteCategory"
-        onClick={() => triggerDeletion(id)}></FaTrashAlt>
+        onClick={() => handleInputMessage('Delete category?')}></FaTrashAlt>
+      {dialog.isLoading && (
+        <ConfirmDialog onDialog={dialogConfirmation} message={dialog.message} />
+      )}
     </div>
   )
 }
